+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - correctly using the parent task when subtask agent creates subtask
 - sorting the list_task tool result tasks by descending updated date
 - added missing gap between messages in virtualized message rendering
+- added support for infinite agent iterations
 
 ## [0.65.0]